filter out team folders. make folders container more modern
