+++ shared/folders/index.desktop.js
@@ -1,32 +1,39 @@
 // @flow
+import Banner from './install/banner'
+import InstallSecurityPrefs from './install/security-prefs'
 import List from './list'
 import React, {Component} from 'react'
-import type {Props} from './render'
-import {Box, TabBar, HeaderHoc} from '../common-adapters'
+import {Box, TabBar} from '../common-adapters'
 import {TabBarItem, TabBarButton} from '../common-adapters/tab-bar'
+import {connect} from 'react-redux'
+import {fuseStatus} from '../actions/kbfs'
 import {globalStyles, globalColors, globalMargins} from '../styles'
-import {compose, defaultProps} from 'recompose'
+import {isLinux} from '../constants/platform'
+
+import type {Props} from '.'
+import type {TypedState} from '../constants/reducer'
 
 class FoldersRender extends Component<Props> {
   _makeItem(isPublic: boolean, isSelected: boolean) {
-    const icon = isPublic ? 'icon-folder-public-24' : 'icon-folder-private-24'
+    const icon = isPublic ? 'iconfont-folder-public' : 'iconfont-folder-private'
+    const selectedColor = isPublic ? globalColors.yellowGreen : globalColors.darkBlue2
+    const iconStyle = isPublic
+      ? {color: globalColors.yellowGreen2, marginBottom: isSelected ? 0 : 0, opacity: isSelected ? 1.0 : 0.6}
+      : {color: globalColors.darkBlue2, marginBottom: isSelected ? 0 : 0, opacity: isSelected ? 1.0 : 0.6}
     return (
       <TabBarButton
         source={{type: 'icon', icon}}
         style={{
           ...styleItem,
-          opacity: isSelected ? 1 : 0.6,
-          borderBottomWidth: 2,
-          borderBottomColor: isSelected
-            ? isPublic ? globalColors.yellowGreen2 : globalColors.darkBlue2
-            : globalColors.transparent,
+          borderBottom: `solid 2px ${isSelected ? selectedColor : 'transparent'}`,
         }}
         styleBadge={styleBadge}
-        styleIcon={styleIcon}
+        styleIcon={{...styleIcon, ...iconStyle}}
         styleLabel={{
           color: isPublic ? globalColors.yellowGreen2 : globalColors.darkBlue,
+          opacity: isSelected ? 1 : 0.6,
+          fontSize: 12,
         }}
-        styleBadgeNumber={styleBadgeNumber}
         selected={isSelected}
         label={isPublic ? 'public/' : 'private/'}
         badgeNumber={isPublic ? this.props.publicBadge : this.props.privateBadge}
@@ -35,20 +42,34 @@ class FoldersRender extends Component<Props> {
   }
 
   render() {
+    if (this.props.showSecurityPrefs) {
+      return <InstallSecurityPrefs />
+    }
+
+    const sharedListProps = {
+      style: this.props.listStyle,
+      smallMode: this.props.smallMode,
+      onRekey: this.props.onRekey,
+      onOpen: this.props.onOpen,
+      onChat: this.props.onChat,
+      onClick: this.props.onClick,
+    }
+
     return (
       <Box
         style={{
           ...stylesContainer,
           backgroundColor: globalColors.white,
+          paddingTop: 0,
+          minHeight: 32,
         }}
       >
+        {!this.props.smallMode && !isLinux && <Banner />}
         <TabBar
           styleTabBar={{
             ...tabBarStyle,
             backgroundColor: globalColors.white,
-            borderBottomWidth: 1,
-            borderBottomColor: globalColors.black_05,
-            marginBottom: globalMargins.xtiny,
+            minHeight: this.props.smallMode ? 32 : 48,
           }}
         >
           {[false, true].map(isPublic => (
@@ -63,8 +84,7 @@ class FoldersRender extends Component<Props> {
             >
               <List
                 {...(isPublic ? this.props.public : this.props.private)}
-                smallMode={this.props.smallMode}
-                onClick={this.props.onClick}
+                {...sharedListProps}
                 isPublic={isPublic}
                 showIgnored={this.props.showingIgnored}
                 onToggleShowIgnored={this.props.onToggleShowIgnored}
@@ -79,45 +99,47 @@ class FoldersRender extends Component<Props> {
 
 const stylesContainer = {
   ...globalStyles.flexBoxColumn,
-  ...globalStyles.fullHeight,
+  flexGrow: 1,
 }
 
 const styleBadge = {
-  borderWidth: 0,
-  paddingLeft: 3,
-  paddingRight: 3,
-  borderRadius: 20,
-  flex: 'initial',
-  justifyContent: 'center',
-  alignItems: 'center',
-  marginRight: 15,
+  marginRight: 2,
   marginLeft: 2,
 }
 
 const styleIcon = {
-  marginRight: globalMargins.xtiny,
+  marginBottom: 2,
 }
 
 const styleItem = {
   ...globalStyles.flexBoxRow,
   paddingTop: 8,
   paddingBottom: 8,
+  paddingLeft: globalMargins.medium,
+  paddingRight: globalMargins.medium,
+  justifyContent: 'flex-start',
   backgroundColor: globalColors.transparent,
-}
-
-const styleBadgeNumber = {
-  lineHeight: '13px',
-  fontSize: 11,
 }
 
 const itemContainerStyle = {
   ...globalStyles.flexBoxColumn,
+  marginBottom: -1,
 }
 
 const tabBarStyle = {
   ...globalStyles.flexBoxRow,
 }
 
-export default compose(defaultProps({title: 'FOLDERS', headerStyle: {borderBottomWidth: 0}}), HeaderHoc)(
-  FoldersRender
-)
+const mapStateToProps = (state: TypedState) => {
+  const installed = state.favorite.fuseStatus && state.favorite.fuseStatus.kextStarted
+  return {
+    installed,
+    showSecurityPrefs: !installed && state.favorite.kextPermissionError,
+  }
+}
+
+const mapDispatchToProps = (dispatch: any) => ({
+  fuseStatus: () => dispatch(fuseStatus()),
+})
+
+export default connect(mapStateToProps, mapDispatchToProps)(FoldersRender)
