+++ shared/folders/index.native.js
@@ -1,38 +1,33 @@
 // @flow
 import List from './list'
 import React, {Component} from 'react'
-import type {Props} from './render'
-import {Box, TabBar} from '../common-adapters'
+import {Box, TabBar, HeaderHoc} from '../common-adapters'
 import {TabBarItem, TabBarButton} from '../common-adapters/tab-bar'
 import {globalStyles, globalColors, globalMargins} from '../styles'
-import {connect} from 'react-redux'
-import {fuseStatus} from '../actions/kbfs'
-import Banner from './install/banner'
-import InstallSecurityPrefs from './install/security-prefs'
-import {isLinux} from '../constants/platform'
-import type {TypedState} from '../constants/reducer'
+import {compose, defaultProps} from 'recompose'
+
+import type {Props} from '.'
 
 class FoldersRender extends Component<Props> {
   _makeItem(isPublic: boolean, isSelected: boolean) {
-    const icon = isPublic ? 'iconfont-folder-public' : 'iconfont-folder-private'
-    const selectedColor = isPublic ? globalColors.yellowGreen : globalColors.darkBlue2
-    const iconStyle = isPublic
-      ? {color: globalColors.yellowGreen2, marginBottom: isSelected ? 0 : 0, opacity: isSelected ? 1.0 : 0.6}
-      : {color: globalColors.darkBlue2, marginBottom: isSelected ? 0 : 0, opacity: isSelected ? 1.0 : 0.6}
+    const icon = isPublic ? 'icon-folder-public-24' : 'icon-folder-private-24'
     return (
       <TabBarButton
         source={{type: 'icon', icon}}
         style={{
           ...styleItem,
-          borderBottom: `solid 2px ${isSelected ? selectedColor : 'transparent'}`,
+          opacity: isSelected ? 1 : 0.6,
+          borderBottomWidth: 2,
+          borderBottomColor: isSelected
+            ? isPublic ? globalColors.yellowGreen2 : globalColors.darkBlue2
+            : globalColors.transparent,
         }}
         styleBadge={styleBadge}
-        styleIcon={{...styleIcon, ...iconStyle}}
+        styleIcon={styleIcon}
         styleLabel={{
           color: isPublic ? globalColors.yellowGreen2 : globalColors.darkBlue,
-          opacity: isSelected ? 1 : 0.6,
-          fontSize: 12,
         }}
+        styleBadgeNumber={styleBadgeNumber}
         selected={isSelected}
         label={isPublic ? 'public/' : 'private/'}
         badgeNumber={isPublic ? this.props.publicBadge : this.props.privateBadge}
@@ -41,34 +36,20 @@ class FoldersRender extends Component<Props> {
   }
 
   render() {
-    if (this.props.showSecurityPrefs) {
-      return <InstallSecurityPrefs />
-    }
-
-    const sharedListProps = {
-      style: this.props.listStyle,
-      smallMode: this.props.smallMode,
-      onRekey: this.props.onRekey,
-      onOpen: this.props.onOpen,
-      onChat: this.props.onChat,
-      onClick: this.props.onClick,
-    }
-
     return (
       <Box
         style={{
           ...stylesContainer,
           backgroundColor: globalColors.white,
-          paddingTop: 0,
-          minHeight: 32,
         }}
       >
-        {!this.props.smallMode && !isLinux && <Banner />}
         <TabBar
           styleTabBar={{
             ...tabBarStyle,
             backgroundColor: globalColors.white,
-            minHeight: this.props.smallMode ? 32 : 48,
+            borderBottomWidth: 1,
+            borderBottomColor: globalColors.black_05,
+            marginBottom: globalMargins.xtiny,
           }}
         >
           {[false, true].map(isPublic => (
@@ -83,7 +64,8 @@ class FoldersRender extends Component<Props> {
             >
               <List
                 {...(isPublic ? this.props.public : this.props.private)}
-                {...sharedListProps}
+                smallMode={this.props.smallMode}
+                onClick={this.props.onClick}
                 isPublic={isPublic}
                 showIgnored={this.props.showingIgnored}
                 onToggleShowIgnored={this.props.onToggleShowIgnored}
@@ -98,47 +80,45 @@ class FoldersRender extends Component<Props> {
 
 const stylesContainer = {
   ...globalStyles.flexBoxColumn,
-  flexGrow: 1,
+  ...globalStyles.fullHeight,
 }
 
 const styleBadge = {
-  marginRight: 2,
+  borderWidth: 0,
+  paddingLeft: 3,
+  paddingRight: 3,
+  borderRadius: 20,
+  flex: 'initial',
+  justifyContent: 'center',
+  alignItems: 'center',
+  marginRight: 15,
   marginLeft: 2,
 }
 
 const styleIcon = {
-  marginBottom: 2,
+  marginRight: globalMargins.xtiny,
 }
 
 const styleItem = {
   ...globalStyles.flexBoxRow,
   paddingTop: 8,
   paddingBottom: 8,
-  paddingLeft: globalMargins.medium,
-  paddingRight: globalMargins.medium,
-  justifyContent: 'flex-start',
   backgroundColor: globalColors.transparent,
+}
+
+const styleBadgeNumber = {
+  lineHeight: '13px',
+  fontSize: 11,
 }
 
 const itemContainerStyle = {
   ...globalStyles.flexBoxColumn,
-  marginBottom: -1,
 }
 
 const tabBarStyle = {
   ...globalStyles.flexBoxRow,
 }
 
-const mapStateToProps = (state: TypedState) => {
-  const installed = state.favorite.fuseStatus && state.favorite.fuseStatus.kextStarted
-  return {
-    installed,
-    showSecurityPrefs: !installed && state.favorite.kextPermissionError,
-  }
-}
-
-const mapDispatchToProps = (dispatch: any) => ({
-  fuseStatus: () => dispatch(fuseStatus()),
-})
-
-export default connect(mapStateToProps, mapDispatchToProps)(FoldersRender)
+export default compose(defaultProps({title: 'FOLDERS', headerStyle: {borderBottomWidth: 0}}), HeaderHoc)(
+  FoldersRender
+)
