+++ shared/menubar/index.render.js.flow
@@ -2,7 +2,7 @@
 import {Component} from 'react'
 
 import type {KBFSStatus} from '../constants/favorite'
-import type {Props as FolderProps} from '../folders/render'
+import type {Props as FolderProps} from '../folders'
 
 export type Props = {
   debug?: boolean,
